twinkleconfig: add new file; twinkle.js: update loader to load new file; twinklespeedy: comment fix
This is a graphical configuration interface for Twinkle's options. To try it out, install the script, and go to [[User:This, that and the other/Twinkle preferences]].
It's hard to write a concise commit message for a new file... All I can say is, look through the code, and/or try it out.

Essentially, there is a hash of all the preferences, which is read by Twinkle.config.init to generate the form interface.
Then Twinkle.config.writePrefs writes the preferences into Twinkle.prefs, which is saved as JSON to the user's twinkleoptions.js subpage.
(The twinkleoptions.js file is loaded by twinkle.js - if it doesn't exist, any TwinkleConfig provided in vector/monobook.js is used, or the default settings are referenced.)

Works in Firefox 4 and Chrome latest. More testing would be welcome.

I copied in Douglas Crockford's implementation of JSON.stringify (public domain code), for older browsers that don't natively implement it.

Things to do:
- The "reset all preferences" code is a bit of a cop-out...
- Implement Friendly prefs
- Fix the XXX comments