all modules: change alert to exception in Twinkle check code
Over 1,300 users still include Twinkle by importing the script filess in vector.js.
This no longer works, due to the niceties of the twinkle.js loader.
The alert was annoying to say the least (I got 10 in a row when testing...)
So we throw an exception, instead, which should land in the browser's error console.