twinkle.js, morebits.js: new loading stuff
This replaces UncleDouggie's loader with a simpler version.
The "load sentry" is a hackish thing which fires a callback when loaded. It is used so we know when twinkleoptions.js is loaded, even if empty.

The .gitignore is included as a courtesy.

I haven't tested this, for fairly obvious reasons. That's why it's in a new branch.