Cleanup: Move everything Twinkle-related from morebits into twinkle.header; Configuration: Load user config before calling Twinkle.load
This is mostly preparations.
Ensuring that user prefs are loaded before calling Twinkle.load allows to move the portlet configuration into twinkleconfig.
In addition, having control over the user prefs allows us to update user config after load, which we will certainly need down the line.
I decided to use parseJSON to read the config options instead of eval since it seemed right to me: We want it to be configurations only and valid JSON, why not enforce it?