twinkleconfig: work around awful display in IE
The preferences panel in IE is about 2500 pixels wide, because of text wrapping issues.
I haven't been able to find a more legitimate fix for this issue... so here is a lovely hack.
I considered simply removing the whiteSpace = "nowrap" rule for IE, but IE's HTML debugger is woeful (the script debugger is fine, but...) so I didn't know if it would work.