warn: Start an additional section for the current month if necessary
Previously Twinkle just looked for a header with the current month and
year, and if it found one, the warning was just appended to the bottom
of the talk page, regardless of whether the current month header was
still the last header on the page.
Now, Twinkle checks whether the last header on the page is titled with
the current month, and if not, a fresh header is added above the
warning.
Fixes #14.