xfd: Patrol pages when nominating for AFD
Also adding a pref to enable/disable this behaviour as required.

The pref is named `markXfdPagesAsPatrolled`, despite only applying to
AFD - this is just in case we ever want to extend it to other XFD
discussion venues. I assume one preference for them all will suffice.

As requested at WT:TW.