speedy: Hide template help when creator is notified
Speedy deletion templates, by default, include the wikitext of the
notification template, which the nominator can copy and paste onto the
page creator's talk page. Since Twinkle notifies the creator
automatically, there is no need for this to be displayed. Twinkle
already passes in "help=off" for XFD, so why not for CSD?

Note that I have only implemented this on-wiki in {{db-test}} so far.