warn: Hack to restore user talk link after rollback
The old, static-page rollback system was replaced by a AJAX version
which no longer displays the confirmation page. Instead, a mw.notify
popup is shown if the rollback is successful.

Unfortunately, there is no good way to hook into the rollback success
logic, so I had to come up with this awful hack. It replaces the
mw.notify function with an override, which adds a link to the talk page
of the user whose edits are being reverted when it detects that a
rollback success message is being displayed.

Hopefully we won't need to keep this around for too long...

See Phabricator tasks T88044 and T136375 for more info.

Also a minor punctuation fix.