Make use of wgRelevantUserName in friendly and fluff
wgRelevantUserName was added to MW (v1.23, https://gerrit.wikimedia.org/g/mediawiki/core/+/7127c1792bb9205d7313ffabb210963c485a2ebc)
in late 2013, well after the friendly modules were brought in to the repo (072860c).
Some uses were later removed as part of a morebits.js tidying (37e9ded),
but wgRelevantUserName is a much cleaner way of getting the username.

The changes in twinklefluff.js are a little more involved, but amount to the same thing.
Here, the over-complicated code for getting the username has been around for even longer,
but the major relevant change was in response to #215.  Basically, the code used to look for a log link
and then calculate the username based off that.  While it obviously still works, wgRelevantUserName
is easier.  #215 dealt with the fact that loading on an empty contribs page ('Special:Contributions/')
broke things, which this still handles via mw.config.exists.
twinklefluff doesn't care if the username doesn't exist, though, but that doesn't seem to cause any problems.