friendlytag: Group old and new redirect tags within Rcat shell
Closes #444.  While processing new tags on a redirect page, check the page for
already-existing redirect categories.  If they exist, whether within or
without an Rcat shell, fold them and the new tags into an Rcat shell.

There is some double-processing, as the already-existing method to check for
duplication remains, but the intent of this is not to affect 'tags' so as not
to affect the edit summary.  Moreover, the old tags remain unsorted, mainly to
avoid creating a broader-scope-than-necessary variable, but is probably a good
idea anyway, to avoid the perception that those tags were added by the user.

Rcat regex courtesy of [[User:Kephir/gadgets/sagittarius.js]] and #433