xfd: Better handling of modulespace/lua at TfD
This needed a lot of tweaks, exacerbated by the tfd/tfm split and the general inflexibility of the related templates.  Honestly, WP:TFD isn't exactly up to snuff and I had to make a number of fixes to templates onwiki as well.
The module's /doc subpage may not exist, so nocreate needed to be changed (added in c97229cd).

A lot of the if/else loops could have been ternaries, but given the length and complexity of some of the items involved I thought it clearer to avoid doing so