Morebits: more documentation (#496)
* morebits: add JSDocs for quickForm, simpleWindow and queryString classes

* morebits: add JSDocs for batchOperation

* morebits: add JSDocs for unbinder; minor docs fixes elsewhere

* morebits: Fix a bug in unbinder: the unbinder fails when the
part to be unbinded (text between prefix and postfix) carries
newlines. I think this is a bug because the starting and closing
nowiki tags, for dealing with which this class is mostly used,
need not always be on the same line. Adding the s flag to the
regex fixes the issue.

* morebits: several docs related fixes