xfd and image: Remove check that prevented loading on history pages
Closes #378

As noted there, the check added in 8b65bb1 meant that twinkleimage would no
longer load on history pages (similar to ad5af2f).  That seems to no longer be
an issue, in particular as the structure and loading of Twinkle has changed
dramatically since then.

These changes are imperfect, however.  In order to allow loading on history
pages, the conditions must be relaxed, such that it is possible for twinkle
image and xfd to load on a local nonredirect page with no file on it.  One
possible workaround could be a (morebits?) test for a file rather than trying
to detect 'mw-imagepage-section-filehistory', but per
[quarry:query/32775](https://quarry.wmflabs.org/query/32775) these seem rather
rare (2-3 a month) so the benefit likely exceeds any perceived costs.