fluff: Add [restore this revision] link to oldid pages
Closes #131

Checking for wgDiffOldId and wgDiffNewId excludes Special:Undelete (where they
are each 0), and so should be sufficient to avoid problematic diffs.

One advantage of the restructuring is that [restore this revision] will now
appear when viewing the first revision as a diff with diff=prev and will not
appear on the left-hand revision when viewing the latest revision with
diff=cur, as may sometimes happen.  Similarly, revert or restore links should
no longer appear on single-revision pages.  This makes some of the
logic/checks a tad involved, as well as the fact that wgDiffNewId and
wgDiffOldId have unbalanced behavior (see [[phab:T214985]]).