block/warn/welcome/xfd: Make sure the preview renderer uses wikitext
Morebits.wiki.preview, specifically beginRender, takes an optional second parameter, defaulting to the current page,
which is used to determine the content model of the output from the parse API query.
We've largely ignored that parameter, meaning that if you attempt to generate a preview while on a non-wikitext page (JS, CSS, JSON, etc.),
beginRender defaults to the model of that page.  This fixes that by explicitly providing a wikitext page.

I chose [[WP:TW]] because it is:
- Already fully protected
- Within the project scope (i.e. "ours")
- Heavily used so unlikely to be changed