speedy/tag: Move histmerge from speedy to tag
{{db-histmerge}} was [renamed](https://en.wikipedia.org/w/index.php?title=Template:History_merge&diff=420055188&oldid=401464558) and [removed](https://en.wikipedia.org/w/index.php?title=Wikipedia%3ACriteria_for_speedy_deletion&diff=prev&oldid=420062064) from the the CSD criteria in March, 2011, following [brief discussion](https://en.wikipedia.org/wiki/Wikipedia_talk:Criteria_for_speedy_deletion/Archive_42#Remove_db-histmerge_from_CSD).  Of interest, the initial import of Twinkle into GitHub happened merely a week earlier.

Both have remained thus for nearly eight years, so we've technically been wrong for quite some time.  `{{history merge}}` is a bit of a weird beast, with a warning associated with it (`{{w-c&pmove}}`) like a speedy; that template is available in `twinklewarn`, and I don't think people expect the tag module to handle those, so I think we're fine not doing that here.  Allows for multiple instances of the template, a la `{{merge from}}`.

Closes #293