friendlytag: Put tags under csd and prod nominations (not afd yet)
Uses regex from the speedy module, partial fix for #237.  AfD is NOT supported because it doesn't lead with a template, but rather with a <!-- comment.

This regex was already enormous and unwieldy, and is now even worse.  It'd be good to clean it up, but even modern js isn't fully pcre, so there's no /x to help us out.