protect: Disable pending changes dropdown on ineligible namespaces
Similar to #451, it would be better if we didn't offer inappropriate options.  The error from the API request is clear enough, but it resets the form.

We *could* turn it off entirely a la #551, but I think it'd be better to do this for 6-12 months before removing it entirely to smooth the transition and get the additional nudge.

In the distant future, it would be nice for this not to be hardcoded, but $wgFlaggedRevsNamespaces isn't exposed to javascript; that's [[[phab:T218479]]](https://phabricator.wikimedia.org/T218479)