fluff: Reload Twinklefluff interface when diff view changes (#407)
This patch will hook Twinklefluff initialization to the mw.hook
responsible for pages containing a diff. Hooking is only done when
the methods for diff pages are loaded.

Now for example the RevisionSlider extension that fires that hook
will re-initiate the Twinkle links added to the page when moving
through the revisions.

This should fix https://github.com/azatoth/twinkle/issues/360

The hook gets fired everytime a diff page loads / reloads. To
avoid having the diff links twice the original method got re-
factored and the part addressing diff pages reacts only to the
hook now.

Further modules that might need to be re-initiated by hook
could be added in a follow up.

Co-authored-by: Amory Meltzer <Amorymeltzer@gmail.com>