block: Remove old fix for indef expirys that is no longer necessary, refactor some regex
Various synonyms for infinite weren't always considered in the blocking api, see https://phabricator.wikimedia.org/T68646.  This was fixed in https://gerrit.wikimedia.org/r/c/mediawiki/core/+/140276 so that infinite, indeinite, infinity, and never are treated the same.  Also allowed for infinity and infinite in the regex matching.