fluff: Add vanarticle parameter to user talk links on diffs
Allows user to click to an editor involved in a diff and provide a warning with the article title filled-in, without having to revert the diff themself (e.g., already-reverted vandalism that went unwarned).  This has been requested a few times, and is fairly trivial.  The link is bolded, just like the user talk link on the mw rollback done page.  Clicking the talkpage link will NOT trigger the autowarn feature (#509) even if enabled.

Also sorted some functions from #562 into the `Twinkle.fluff` object.