warn: Reorganize messages object, should make it more clear to maintain
Setting it up this way means there is even more of a difference between the scaled levels (level2, level4im, etc.) and singlenotice/singlewarning/custom,
but doing so should make it very straightforward to remove or add templates (pursuant to any discussion in #533).

The only change to the enduser is that this will now generate a submenu category for every grouping, even if there are no templates in that group at that level.  At the moment, this is only the `4im` level for the "Removal of deletion tags" category.  I'm not sure this is a bad thing.

One header was mistmatched ('other editors' versus 'other users'), resolves that.