warn: remove old fixes for out-of-range defaultGroup
This dates back to 2008 (<https://en.wikipedia.org/w/index.php?diff=247182138&oldid=245908598&title=User:AzaToth/twinklewarn.js>).  There used to be a group 8 with block templates for sysops, hence the `Morebits.userIsInGroup` switch, but after #260 it was removed and this persisted, meaning that we don't need any special handling for sysops or nonsysops here.  That is, `defaultGroup > 7` would handle any remaining unmodified options (there are some).  Similarly, undefined or -1 groups were possible due to how the config was structured back then: in your skin.js rather than a separate file.

Perhaps old browsers threw an error, but for now I can't seem to trigger any errors and regardless, the default behavior if something isn't true is the first option, so none of this is needed.