hide TalkList in ns 447 (Education Program talk)
There is not a single page in namespace 446 (Education Program), so all such pages are talk pages of non-existent pages. But, they continue to exist because, per w.wiki/3Kv, "However, the namespace was re-added later that year to allow access to old talk pages in the Education Program talk: namespace." Thus, any pages in the 447 namespace falls under the exclusion of "any page that is useful to Wikipedia".