Rephrase DI notice
Since F4, F5, F6, and F11 are hidden from non-sysops, the warning shouldn't say "can be done using Twinkle's "DI" tab" but rather "is done using...", since can implies that the DI tab is in addition to the CSD tab. "Can only" is another option, but it implies that tools other than Twinkle cannot be used.