Add redirnone to categoryList
Currently, db-redirnone ("G8: Redirects to invalid targets...") is only shown on pages that Twinkle detects are redirects. However, per w.wiki/3L4 categories are generally not redirected using "hard redirects" but rather using the "Category redirect" template. Rather than moving this instance of G8 to the general list, and requiring that it be shown only on redirects unless in the category namespace, or changing how redirects are detected, I just added it to the category list.