fluff: Remove overly parsimonious check for diffs, blocked links if diff content hidden
Fixes #627.  If a user has "Do not show page content below diffs" selected in Special:Preferences, `wgRevisionId` will be set to 0 on diffs, meaning no links would be shown on diffs.  Comparing `wgRevisionId` to `wgDiffNewId` from #563 appears to be a holdover from some (local) attempts to handle merge conflicts between #511 and #407, when I initially had this structured differently, but that should now handled in the init, so this was excessive.