fluff: don't create rollback links for single-revision pages on contribs
Possibly a result from the fix in #467, but either way, this would create rollback/vandal links for every page marked with `mw-uctop`, even if it had only a single revision and thus couldn't be reverted.  The rollback links would fail (using & instead of ?) but wouldn't have been properly constructed anyway.  Check for the existence of a `href` here should remove those links.