morebits: overhaul action complete status messages
The action complete status messages (the last line of status)
in almost all Twinkle modules do not read correctly. Such as in Tag
module, it says "Tagging complete: completed". In xfd, it says
"Nomination complete, redirecting to discussion page: completed".
None of the Twinkle modules make meaningful use of Morebits.wiki.
actionCompleted.postfix (the part of the status message after the
colon), allowing it to remain "completed" which is the default.
There is no way to suppress the display of the colon and postfix.

This commit introduces a new function that creates a status line
without a colon separator, and uses it for the action-completed
message instead of using a normal Morebits.status object. This
does away with Morebits.wiki.actionCompleted.postfix, which as
described above, has no utility.

The default value of Morebits.wiki.actionCompleted.notice is changed
from "Action", which does not make much sense, to null. If it is not
redefined in the module, the entire action-complete message is suppressed.
This was not possible earlier, and was being achieved in RPP/PP through
an ugly hack.