morebits: quickform: textarea: nest a label within the h5
This was marked as a TODO item in source. The advantage is that now clicking on the label will put the textarea in focus.

Also mark the textarea as the childContainer, so that id, classes and styles given are applied to the textarea rather than to the container div. This is desirable in most situations.