morebits: quickform: input: fix critical ancient bug
Until this commit, the input (text) element and its parent div element were getting the same id, which is invalid html.
As a result of this, clicking on the label, despite it having the `for` attribute, wouldn't cause its corresponding input field to get focus.

As the subnode (the input element) is now the `childContainer`, any id, classes and styles given are applied to this element rather than to the parent div.