warn: Add option to show a list with all template warnings
Mentioned by me in https://github.com/azatoth/twinkle/issues/673#issuecomment-501858213.  While this makes use of `defaultWarningGroup = 10`, I considered reusing 8, which was until the addition of the block module in #260 (well, also 430e2258) reserved for sysops to select blocking templates, to cut down on bloat.  There are a handful of users, some still active sysops, with the option set, but while that's a trivial barrier, I think it's better list it below the individual items, reinforcing the "all-of-the-above"-ness of the option, as well as not relying other projects to do the same thing.  Unlike the combined singlet menu from #759, this menu is not sorted: I thought it'd be clearer for folks to have a progression rather than all the groups clumped together, to take advantage of the quickFilter.

This also tweaks the process by which `twinklewarn` tries to select the template that corresponds to what was selected in the previous menu.  As it contains every item, this new `kitchensink` menu should always match exactly, but the previous process kept searching once a match was found.  This wasn't an issue since this was only relevant for levels 1-4im where only the correct match could be found, but here it meant that the highest level warning would always be selected upon moving to the kitchensink menu (i.e., vandalism1 -> vandalism4im).  Now, in addition to doing an exact match in the kitchensink, it also won't keep searching once it has found the match.  Part of this is as a result of my work in #592.  There's also a tweak to how the combined singlet menu from #759 is sorted; it now selects either the previously selected option (if going from kitchensink->singletcombined, just like for other menu changes), or the top option if the previous value is not present (level2->singletcombined, as in #759).

I also removed a comment made incorrect by #641.