warn: Add option to autoselect warning level 1-4
Dynamically loads the correct level as determined based on current warnings, in preview and to generate the autolevel menu itself.

The autolevel decision means the 24-hour identical check isn't necessary, since if something is detected, autolevel won't ever generate an identical template.  This uses option 11 in the `defaultWarningGroup` config.  Would be nice to have it underneath 4im but for the sake of ordinality, this seems better.  Adds stale length as a hidden preference (for now, but could be made user customizable), defaulting `autolevelStaleDays` to 3 days.  Comparison to <s>our competitors</s> other tools:

CBNG resets after 2 days
     https://github.com/DamianZaremba/cluebotng/blob/4958e25d6874cba01c75f11debd2e511fd5a2ce5/bot/action_functions.php#L62
Huggle defaults to 30
       https://github.com/huggle/huggle3-qt-lx/blob/13df16adf5744825f8a27d70b3bab998ad19f4b0/src/huggle_core/configuration.cpp#L187
       https://github.com/huggle/huggle3-qt-lx/blob/13df16adf5744825f8a27d70b3bab998ad19f4b0/src/huggle_core/projectconfiguration.hpp#L172
       On enwiki, Huggle is now (and previously) 3 days: https://en.wikipedia.org/w/index.php?title=Wikipedia:Huggle/Config.yaml&diff=919417999&oldid=918243346

Makes use of `Twinkle.warn.talkpageObj` to hold on to the page object whenever it is loaded which is: the first time the autolevel is selected, every preview, and in the final evaluation.  The date processing stuff from the main evaluation is now pulled out into a separate function.  If the last template found doesn't have a level, this checks `Twinkle.warn.messages.singlewarn` (which it treats as level 3) before defaulting to level 1.

Requiring the page object meant duplicating the createEntries loop, but alas.  Likewise, there was a need to only handle select2 and other submenu stuff *after* the load, so I've put that in a function and called it for autolevel and separately below for everyone except autolevel.

There's a new autolevel message added/removed as appropriate alerting the user to the level in question (it's show in the menu, but that's easy to miss).  Additionally, if the autolevel detects a prior level 4 (or 4im) warning, , the menu will generate a clickable AIV link in-line, similar to what speedy does for unlink.  Supplies the relevant page, if any, and checks the "vandalism after final" box.  Some messages have been tweaked, and I also removed a duplicate `tw-warn-red-notice removal`: added in fb92245c, unneeded since covered in `change_subcategory`.