morebits.wiki.page: save: pass the api object to the failure handler
Presently, the morebits.wiki.page object is passed to the failure handler.
But this object is of no use as it contains no information about the error.
The morebits.wiki.api object is more useful as it carries the errorCode and
errorText properties, and can thus be used to determine the error and produce
a customised error message.

Also, the page object still remains accessible as it is the parent attribute
of the api object. (While the api object is not accessible from the page object.)

The success callback continues to take the page object as parameter. This is
probably desired for the success callback to allow additional processing with
the page after save.

This change does not break anything in Twinkle.