morebits.wiki.page: deletePage: ensure consistent parameters in callbacks
Give the apiobj as the parameter to the failure handler on all types of errors (except, of course,
when the error is handled before any API calls take place), rather
than doing the mix of giving the pageobj on some errors and apiobj on some errors.

L2856 - `ctx.onDeleteFailure.call(this, this, ctx.deleteProcessApi);` was most probably a typo, as
it passes two arguments to the fail callback, which is not done anywhere else.

Remove `if (ctx.onDeleteFailure)` checks since these will always evaluate to true, as
`ctx.onDelelteFailure` fallbacks to an empty function (L2301) if not given by user.