morebits: Update handling of blacklist and abusefilter errors from API (breaking change for 1.34.0-wmf.23)
Closes #704

As [recently announced](https://lists.wikimedia.org/pipermail/mediawiki-api-announce/2019-September/000148.html), a [breaking change](https://phabricator.wikimedia.org/T229539) is being made to how the API returns errors for the abuse filter and the spam blacklist; the differences can be seen in [T232583](https://phabricator.wikimedia.org/T232583), but the tl;dr is that instead of reporting a successful edit as "failure," these errors will finally (as of 1.34.0-wmf.23) be correctly reported as error codes.

This is, obviously, good, and simplifies things quite a bit on our end, in particular handling the two abusefilter option (`abusefilter-disallowed` and `abusefilter-warning`) in a reasonable fashion.  These need slightly different messages, and the full `info` response is a little too long.  On the other hand, while the default system response to `spamblacklist` is wordy, it seems to cover all the bases without being too excessive.  We're not currently retrying `abusefilter-warning` so I've folded all of these into the "hard error" section.

As for `twinklefluff.js`, the `Morebits.wiki.api` call to `Twinkle.fluff.callbacks.complete` doesn't include an `onError` call, so unless I'm mistaken, we won't need to handle them here anymore.  Theoretically, this should marginally help with #182.