twinkle: add menus for monobook and modern skins
Add menus for monobook and modern skins. Code adapted from that of [[User:Haza-w/Drop-down menus]] at [[MediaWiki:Gadget-dropdown-menus-nonvector.js]] and [[MediaWiki:Gadget-dropdown-menus-nonvector.css]], with modifications.

Since the code for generation of menus in monobook/modern does not have anything in common with the code for creating menu in vector or buttons in nonvector, this has been put wholly in an `if` block.