twinkle: fix addPortletLink for `Last` button in monobook/modern menu
If the user clicks on the `li` element (but not within the `a` element), trigger the clicking of the `a` anyway.
This is useful for the `Last` button (the only one that leads to a page than a function execution) in the newly added menus, as here when the cursor is hovering over the `li`, the underline appears under the text, but without this clicking would have no effect.