xfd: add support for RM (#698)
Adds the ability to start an RM discussion on the talk page, and to list uncontroversial requests at [[WP:RM/TR]].
Closes #644.

While this is admittedly not a best fit in XFD module as it is not an XFD board, it is great to have this feature somewhere in Twinkle. This provides an easy interface to access the RM process, as users now no longer need to remember the usage of the templates involved. Also, this would probably help new editors to discover the RM process sooner.