welcome: put all template and grouping data in a single object
Tweaked the structure of Twinkle.welcome.templates so that it contains all the information regarding what templates are to shown under what headers, and in which modes. This makes i18n/l10n lot easier as now only this object needs to redefined for different wikis (apart from the translation of strings). Also this makes maintenance easier as the only code change needed is in this object while adding/removing templates.