docs: Load morebits as a gadget (#750)
It was [suggested at [[WP:IANB]]](http://en.wikipedia.org/wiki/Special:Permalink/926710304#Hidden_gadgets) that it would be helpful to load morebits as a hidden gadget in [[MediaWiki:Gadgets-definition]].  Making the morebits library available like this would be helpful for authors and users of other tools, and of course makes sense as it is intended to serve as a library.  This doesn't require any *code* changes on our part, but the documentation should be updated to reflect current practice for maintainers.

Closes #748.