morebits: batchOperation: improve status line generation
If using mw.Api() (or anything other than Morebits.wiki.page/api) for processing pages in the worker, till now it was not possible to have invidual status lines for each page. This commit rectifies this deficiency by allowing a string input to workerSuccess() that would be treated as the page name, triggering individual status line generation.

I've changed the condition `if (apiobj && apiobj.getStatusElement)` to `if (apiobj instanceof Morebits.wiki.page || apiobj instanceof Morebits.wiki.api)`. These are functionally same, but the latter is more explicit and clearer.