warn: Note if user attempting to warn for an old edit
This will show a message in red to the user if they are considering issuing a warning older than 24 hours, and may thus be stale.  This is something that Huggle already does, with a [limit of 1 day](https://github.com/huggle/huggle3-qt-lx/blob/f44d4f7ea0f7b83fb16e4d14e9bedf3c18b09f15/src/huggle_core/warnings.cpp#L73) for the alert.  24 hours is also the window in which Twinkle warns users that they are attempting to provide an identical warning to one already present, so the parallel "staleness" is nice.

We can do this by using the `vanarticlerevid` from `twinklefluff` on diffs to query for the timestamp; ideally, diff timestamps would be available without an API call (T239439).  The rollback success page exposes neither the revision or time of the edit in question, nor does the url, so we can't check for old diff from the user talk link; this is comparatively rare anyway.

Also use .getTime() throughout when comparing date objects.