Add new option to disable individual Twinkle modules
Not everybody uses everything, and even in a menu, a lot of menus you never intend on using is annoying.  This adds two new config options, `Twinkle.config.disabledModules` and `Twinkle.config.disabledSysopModules`, and constructs an array of module names to be turned off.  The loading mechanism in `Twinkle.js` now only loads items not explicitly turned off.  The instances where we link from one to another (deli->speedy, speedy->unlink) work fine since they use the `callback`.  `config` cannot be turned off, and the default is, of course, to have everything on.