arv: Remove moment dependency
See #769 for the rationale, but tl;dr the only uses of [momentjs](https://momentjs.com/) are in `twinklearv` and are cosmetic, so loading moment is wasteful.  There are two functions we need to replace, `.format()` and `.calendar()`; this uses `.toLocaleString()` and `.toLocaleDateString()`, respectively.

Regarding formatting, doing:

	var localeOptions = {
	      hour: 'numeric',
	      hour12: false,
	      minute: 'numeric',
	      day: 'numeric',
	      month: 'long',
	      year: 'numeric',
	      timeZone: 'UTC',
	      timeZoneName: 'short'
	};
	new Date().toLocaleString(undefined, localeOptions)

produces:
November 25, 2019, 21:36 UTC

This doesn't fully recpitulate `moment().utc().format('HH:mm, D MMMM YYYY [(UTC)]')`, which produces:
21:36, 25 November 2019 (UTC)

The difference is negligible, and although it was chosen (0beacc07) to match on-wiki formats, that's a user preference and, notably, *not* a valid date format (see #757).  A more exact reproduction can be achieved by combining some strings, doing something like:

    	var localeTimeOptions = {
		timeStyle: 'short',
		hour12: false,
		timeZone: 'UTC'
	};
	var d = new Date();
	d.toLocaleTimeString(undefined, localeTimeOptions) + ', ' + d.toLocaleDateString(undefined, {day: 'numeric', timeZone: 'UTC'}) + ' ' + new Date().toLocaleDateString(undefined, {month: 'long', year: 'numeric', timeZone: 'UTC'}) + ' (UTC)';

We could place this in a function, but hardly seems necessary, especially since, depending on their selected user preference, not all users are used to the same formats.

As for the relative dates from the `.calendar()` function, they simply can't be easily replaced.  This implementation uses `.toLocaleDateString()` to produce `Fri, Nov 22, 2019, 2:03 PM`, whereas `.calendar()` would have produced `Last Friday at 2:03 PM`.  The short month and weekday names are used to declutter the window as compared to a long string.